+++ app/src/main/java/com/ayush/weatherapp/repository/forecast/ForecastRepository.java
@@ -1,9 +1,9 @@
-package com.ayush.weatherapp.repository.weather;
+package com.ayush.weatherapp.repository.forecast;
 
 import com.ayush.weatherapp.entities.forecast.ForecastEntity;
 import io.reactivex.Observable;
 
-public interface WeatherRepository {
+public interface ForecastRepository {
 
   Observable<ForecastEntity> getForecast(double lat, double lng, boolean isCurrentLocation);
 }
